metaMLST_functions.py edited online with Bitbucket
Former-commit-id: c9cf619f61683bb0ab4dbb8542e0a0501560c530

+++ metaMLST_functions.py
@@ -121,23 +121,22 @@ def stringDiff(s1,s2):
 		if a!=b: c+=1
 	return c 
 
-def buildConsensus(bamFile,chromosomeList,filterScore,max_xM,debugMode):
+def buildConsensus(bamFile,chromosomeList,filterScore,max_xM,debugMode,presorted=False):
 
  	if pkgutil.find_loader('pysam') is not None:
  		import pysam
- 		print "PYSAM"
 	else:
-		print "LEGACY"	
 		return buildConsensus_legacy(bamFile,chromosomeList,filterScore,max_xM,debugMode)
 
 
 	seqRec = []
  	chromosomes = {}
 
-	subprocess.call(['samtools','sort',bamFile,'-o',bamFile+'.sorted'])
-	pysam.index(bamFile+'.sorted')
-	
-	samfile = pysam.AlignmentFile(bamFile+'.sorted', "rb")
+	if not presorted:
+		subprocess.call(['samtools','sort',bamFile,'-o',bamFile+'.sorted'])
+		pysam.index(bamFile+'.sorted')
+		samfile = pysam.AlignmentFile(bamFile+'.sorted', "rb")
+	else: samfile = pysam.AlignmentFile(bamFile, "rb")
 	chromosomesLen = dict((r,l) for r,l in zip(samfile.references,samfile.lengths))
 
 	for chre in chromosomeList.keys():
@@ -200,7 +199,7 @@ def buildConsensus(bamFile,chromosomeList,filterScore,max_xM,debugMode):
 
 
 	samfile.close()
-	os.unlink(bamFile+'.sorted')
+	if not presorted: os.unlink(bamFile+'.sorted')
 	#seqRec.append(SeqRecord(Seq(sequen,IUPAC.unambiguous_dna),id=chromo, description = 'CI::'+str(cIndex)+'_SP::'+str(SNPs)))
 	
 	return seqRec
